# swagger.json
{
	"components": {
...
		"schemas": {
...
			"EConversionStatus": {
				"enum": [
					"converted",
					"in queue",
					"processing"
				],
				"type": "string"
			},
			"IConversionStatus": {
				"properties": {
					"conversionId": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/EConversionStatus"
					}
				},
				"required": [
					"conversionId",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ffmpeg-webservice",
		"version": "0.2.0",
		"description": "REST api to convert files to different formats using ffmpeg",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Sven Patrick Meier ",
			"email": "sven.patrick.meier@team-parallax.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/conversion/{fileId}": {
			"get": {
				"operationId": "GetConvertedFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IConversionStatus"
								}
							}
						}
					}
				},
				"description": "Returns the current status for a conversion given a conversionId",
				"tags": [
					"Conversion"
				],
				"security": [],
				"parameters": [
					{
						"description": "Unique identifier for the conversion of a file.",
						"in": "path",
						"name": "fileId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/"
		}
	]
}